networkutil: read the whole endpoint builder set against arm64
Every format string and endpoint path here was decoded from its
__cfstring entry rather than taken from a decompiler label, and every
stringWithFormat: argument list was recovered from the stp/str writes
to [sp, #..] ahead of the bl. That turned up three defects the source
could not show on its own.

+unlockMusicURL:randKey: dropped an argument. The format at 0x363de0
is target=%@&music=%d&key=%d&%@, and 0x330b8 and 0x330bc fill four
slots: region, music, key, and userInfo. The reconstruction had a
three-slot format and never sent the device fingerprint, so the query
was short one parameter and the trailing specifier had nothing to
consume.

The fingerprint seed suffix was wrong. The __cfstring at 0x363ce0
points at 0x32f6a8 with length 5, which is STORE, not _STORE. A
leading underscore would have changed every MD5 the app sends.

+createNonce: took an int. The type encoding is @24@0:8Q16 and the
disassembly agrees throughout: the cbz at 0x32634 tests x20, and the
loop compares x24 against x20 with b.cc at 0x326bc, all 64-bit and
unsigned. It is now NSUInteger, which also lets the StoreUtil forwarder
pass its length through instead of narrowing it to int first.

The remaining 28 routines matched as written and are ticked with the
addresses that prove them.

+++ Project/NetworkUtil.h
@@ -81,7 +81,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  * @brief The music-unlock endpoint for a given music id, echoing @p randKey to guard against a
- * stale response.
+ * stale response and carrying the device user-info parameters.
  * @param musicID The music identifier to unlock.
  * @param randKey The random key echoed back by the response.
  * @ghidraAddress 0x33058
@@ -99,7 +99,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @return The generated nonce string.
  * @ghidraAddress 0x32610
  */
-+ (nullable NSString *)createNonce:(int)length;
++ (nullable NSString *)createNonce:(NSUInteger)length;
 
 /**
  * @brief The startup / web-info endpoint (@c .../akx/main/cgi/startup/), carrying the region code.
